refactor: Create helpers  file to ContactForm
...and move all helpers functions in `contactToFormValues` & `formValuesToContact`
