+++ src/components/ContactCard/ContactForm/contactToFormValues.spec.js
@@ -1,7 +1,4 @@
-import contactToFormValues, {
-  moveToHead,
-  makeItemLabel
-} from './contactToFormValues'
+import contactToFormValues from './contactToFormValues'
 import { johnDoeContact, johnDoeFormValues } from '../../../helpers/testData'
 
 describe('contactToFormValues function', () => {
@@ -123,47 +120,5 @@ describe('contactToFormValues function', () => {
 
     const result = contactToFormValues(contact, tSpy)
     expect(result).toEqual(expected)
-  })
-})
-
-describe('moveToHead function', () => {
-  it('should move an item to head of the array', () => {
-    const items = [1, 5, 657, 42, 3, 27, 88, 3, 4]
-    const shouldBeHead = v => v === 42
-    const expected = [42, 1, 5, 657, 3, 27, 88, 3, 4]
-    const actual = moveToHead(shouldBeHead)(items)
-    expect(actual).toEqual(expected)
-  })
-})
-
-describe('makeItemLabel', () => {
-  it('should return undefined if no arg', () => {
-    const res = makeItemLabel()
-
-    expect(res).toBe(undefined)
-  })
-
-  it('should return undefined if nothing defined', () => {
-    const res = makeItemLabel({ type: undefined, label: undefined })
-
-    expect(res).toBe(undefined)
-  })
-
-  it('should return correct type and label', () => {
-    const res = makeItemLabel({ type: 'cell', label: 'work' })
-
-    expect(res).toBe('{"type":"cell","label":"work"}')
-  })
-
-  it('should return only label if no type', () => {
-    const res = makeItemLabel({ type: undefined, label: 'work' })
-
-    expect(res).toBe('{"label":"work"}')
-  })
-
-  it('should return only type if no label', () => {
-    const res = makeItemLabel({ type: 'cell', label: undefined })
-
-    expect(res).toBe('{"type":"cell"}')
   })
 })
