+++ src/components/ContactCard/ContactForm/formValuesToContact.js
@@ -1,88 +1,5 @@
-import isEqual from 'lodash/isEqual'
-
-import contactToFormValues from './contactToFormValues'
+import { createAddress, makeTypeAndLabel } from './helpers'
 import { updateIndexFullNameAndDisplayName } from '../../../helpers/contacts'
-
-const hasExtendedAddress = addressField => {
-  if (!addressField) return false
-  const extendedAddressKeys = [
-    'addresslocality',
-    'addressbuilding',
-    'addressstairs',
-    'addressfloor',
-    'addressapartment',
-    'addressentrycode'
-  ]
-  return Object.keys(addressField).some(ext =>
-    extendedAddressKeys.includes(ext)
-  )
-}
-
-/**
- *
- * @param {string} [itemLabel] - Value of the label for a contact attribute
- * @returns {{ type?: string, label?: string }}
- */
-export const makeTypeAndLabel = itemLabel => {
-  if (!itemLabel) {
-    return { type: undefined, label: undefined }
-  }
-
-  const itemLabelObj = JSON.parse(itemLabel)
-
-  const res = { type: itemLabelObj.type, label: itemLabelObj.label }
-
-  return res
-}
-
-const createAddress = ({ address, oldContact, t }) => {
-  return address
-    ? address
-        .filter(val => val && val.address)
-        .map((addressField, index) => {
-          const oldContactAddress = oldContact?.address?.[index]
-          const oldContactFormValues = contactToFormValues(oldContact, t)
-            ?.address?.[index]
-
-          const addressHasBeenModified = !isEqual(
-            addressField,
-            oldContactFormValues
-          )
-
-          if (addressHasBeenModified) {
-            // Use "code" instead "postcode", to be vcard 4.0 rfc 6350 compliant
-            // eslint-disable-next-line no-unused-vars
-            const { postcode, ...oldContactAddressCleaned } =
-              oldContactAddress || {}
-            return {
-              // For keep other properties form connectors
-              ...oldContactAddressCleaned,
-              formattedAddress: addressField.address,
-              number: addressField.addressnumber,
-              street: addressField.addressstreet,
-              code: addressField.addresscode,
-              city: addressField.addresscity,
-              region: addressField.addressregion,
-              country: addressField.addresscountry,
-              ...(hasExtendedAddress(addressField) && {
-                extendedAddress: {
-                  ...oldContactAddressCleaned.extendedAddress,
-                  locality: addressField.addresslocality,
-                  building: addressField.addressbuilding,
-                  stairs: addressField.addressstairs,
-                  floor: addressField.addressfloor,
-                  apartment: addressField.addressapartment,
-                  entrycode: addressField.addressentrycode
-                }
-              }),
-              ...makeTypeAndLabel(addressField.addressLabel),
-              primary: index === 0
-            }
-          }
-          return oldContactAddress
-        })
-    : []
-}
 
 const formValuesToContact = ({ formValues, oldContact, t }) => {
   const {
